@@ -3,14 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^4.3.1",
-    "react": "^16.9.0",
-    "react-bootstrap": "^1.0.0-beta.6",
-    "react-dom": "^16.9.0",
-    "react-mdl": "^1.11.0",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "3.1.1",
-    "styled-components": "^4.3.2"
+    "bootstrap": "^5.3.2",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.9.1",
+    "react-dom": "^18.2.0",
+    "react-mdl": "^2.1.0",
+    "react-router-dom": "^6.18.0",
+    "react-scripts": "5.0.1",
+    "styled-components": "^6.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
