@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Home from './Home';
 import Resume from './Resume';
 import Contact from './Contact';
@@ -12,16 +12,16 @@ class App extends Component {
   render() {
     return (
       <>
-            <Router>
-              <NavBar />
-              <Jumbotron />
-              <Switch>
-                <Route exact path="/" component={Home} />
-                <Route path="/resume" component={Resume} />
-                <Route path="/contact" component={Contact} />
-                <Route component={NoMatch} />
-              </Switch>
-            </Router>
+        <BrowserRouter>
+          <Jumbotron />
+          <NavBar />
+          <Routes>
+            <Route exact path="/" component={Home} />
+            <Route path="/resume" component={Resume} />
+            <Route path="/contact" component={Contact} />
+            <Route component={NoMatch} />
+          </Routes>
+        </BrowserRouter>
       </>
     );
   }
